pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-data-pipeline"
version = "0.1.0"
description = "AI-powered financial data pipeline: ingest, process, and analyze financial news and SEC filings"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Manay Divatia"}]
keywords = ["finance", "kafka", "streaming", "llm", "data-pipeline", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.3,<3",
    "faust-streaming>=0.11,<1",
    "aiokafka>=0.12,<0.13",
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.1,<3",
    "sqlalchemy>=2.0,<3",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "boto3>=1.34",
    "pyarrow>=15.0",
    "httpx>=0.27",
    "feedparser>=6.0",
    "tenacity>=8.2",
    "structlog>=24.1",
    "orjson>=3.9",
    "faker>=24.0",
    "openai>=1.12",
    "anthropic>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "testcontainers[kafka,postgres,minio]>=4.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[tool.setuptools.packages.find]
include = ["src*", "infra*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print (warn on print statements)
]
ignore = ["T201"]  # allow print() in consumer/scripts

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
